[refact]: 清除warning 重构utils
