--- src/git/basic/mod.rs
@@ -1,2 +1,0 @@
-//!打算把commit tag tree blob都放在这个模块下方便一些
-//
